+++ tests/test_sync_dirs.py
@@ -1,7 +1,5 @@
 __author__      = "Jérôme Cuq"
-__copyright__   = "Copyright 2026, Jérôme Cuq"
 __license__     = "BSD-3-Clause"
-
 
 from helpers import log
 from dirsyncer import *